# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
        "/api/dashboard/kpis": {
            "get": {
                "tags": [
                    "Dashboard & Reports"
                ],
                "summary": "Get main Key Performance Indicators (KPIs)",
                "operationId": "e3ffaa45376734cb46f87e051250b1f5",
                "responses": {
                    "200": {
                        "description": "KPIs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "KPIs retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_revenue": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 125000.5
                                                },
                                                "total_orders": {
                                                    "type": "integer",
                                                    "example": 520
                                                },
                                                "new_orders_today": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "active_users": {
                                                    "type": "integer",
                                                    "example": 300
                                                },
                                                "delivery_rate": {
                                                    "description": "Percentage of delivered orders",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.92
                                                },
                                                "delivered_orders": {
                                                    "description": "Total delivered orders count",
                                                    "type": "integer",
                                                    "example": 478
                                                },
                                                "orders_ready_to_ship": {
                                                    "description": "Nombre de commandes en attente d'expédition.",
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "deliveries_in_progress": {
                                                    "description": "Nombre de commandes avec le statut IN_DELIVERY.",
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "deliveries_successful": {
                                                    "description": "Nombre total de livraisons réussies.",
                                                    "type": "integer",
                                                    "example": 478
                                                },
                                                "deliveries_failed": {
                                                    "description": "Nombre total de livraisons échouées.",
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/sales-over-time": {
            "get": {
                "tags": [
                    "Dashboard & Reports"
                ],
                "summary": "Get sales volume and revenue trend over a period (e.g., last 30 days)",
                "operationId": "be25a5f67ebff827e0969f2cd3450be7",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Granularity of aggregation",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "day",
                                "week",
                                "month"
                            ]
                        },
                        "example": "day"
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to look back",
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "label": {
                                                        "type": "string",
                                                        "example": "2025-12-01"
                                                    },
                                                    "revenue": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500.5
                                                    },
                                                    "order_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/top-products": {
            "get": {
                "tags": [
                    "Dashboard & Reports"
                ],
                "summary": "Get top selling products (quantity and revenue)",
                "operationId": "451a834220d0866a8b323f4adc395400",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_name": {
                                                        "type": "string"
                                                    },
                                                    "total_quantity": {
                                                        "type": "integer"
                                                    },
                                                    "total_revenue": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/order-status-distribution": {
            "get": {
                "tags": [
                    "Dashboard & Reports"
                ],
                "summary": "Get the count of orders by status for pie charts",
                "operationId": "b4de423c6cecc86e5a1abb2e8a3e181a",
                "responses": {
                    "200": {
                        "description": "Status distribution retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "PENDING": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "DELIVERED": {
                                                    "type": "integer",
                                                    "example": 400
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}